d temp commit 3

+++ app/src/main/java/com/udtech/drills/feature/content/views/IContentFragmentView.java
@@ -23,5 +23,5 @@ import java.util.List;
 
   void fillCalendar(List<Integer> integerList);
 
-  void setTotalTimesPerWeek(List<Integer> integerList);
+  void setTotalTimesPerWeek(List<Double> integerList);
 }
